% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{shinymde_server}
\alias{shinymde_server}
\title{Server ("backend") for shinymde}
\usage{
shinymde_server(input, output, session)
}
\arguments{
\item{input}{input as defined in the ui}

\item{output}{output of a defined ui tag.}

\item{session}{Session under which the app is running.}
}
\value{
Server for the user interface.
}
\description{
Server ("backend") for shinymde
}
