% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{shinymde_ui}
\alias{shinymde_ui}
\title{Builds the user interface for shinymde.}
\format{
An object of class \code{shiny.tag} of length 3.
}
\usage{
shinymde_ui
}
\value{
shinymde's user interface.
}
\description{
Builds the user interface for shinymde.
}
\keyword{datasets}
